requirements.txt
faker==20.1.0
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2
pydantic==2.5.2
sqlalchemy>=2.0.0
psycopg2-binary
neo4j
python-dotenv
